LayoutTests/imported/w3c/web-platform-tests/fetch/metadata/generated/css-images.sub.tentative-expected.txt

Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY
Blocked access to external URL http://www.localhost:8800/fetch/metadata/resources/record-headers.py?key=GENERATED_KEY


Harness Error (TIMEOUT), message = null

TIMEOUT background-image sec-fetch-site - Not sent to non-trustworthy same-origin destination Test timed out
FAIL border-image sec-fetch-site - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-site - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-site - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-site - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-site - Not sent to non-trustworthy same-site destination Test timed out
FAIL border-image sec-fetch-site - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-site - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-site - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-site - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-site - Not sent to non-trustworthy cross-site destination Test timed out
FAIL border-image sec-fetch-site - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-site - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-site - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-site - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-mode - Not sent to non-trustworthy same-origin destination Test timed out
FAIL border-image sec-fetch-mode - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-mode - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-mode - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-mode - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-mode - Not sent to non-trustworthy same-site destination Test timed out
FAIL border-image sec-fetch-mode - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-mode - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-mode - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-mode - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-mode - Not sent to non-trustworthy cross-site destination Test timed out
FAIL border-image sec-fetch-mode - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-mode - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-mode - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-mode - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-dest - Not sent to non-trustworthy same-origin destination Test timed out
FAIL border-image sec-fetch-dest - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-dest - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-dest - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-dest - Not sent to non-trustworthy same-origin destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-dest - Not sent to non-trustworthy same-site destination Test timed out
FAIL border-image sec-fetch-dest - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-dest - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-dest - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-dest - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-dest - Not sent to non-trustworthy cross-site destination Test timed out
FAIL border-image sec-fetch-dest - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-dest - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-dest - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-dest - Not sent to non-trustworthy cross-site destination assert_unreached: Reached unreachable code
PASS background-image sec-fetch-user - Not sent to non-trustworthy same-origin destination
PASS border-image sec-fetch-user - Not sent to non-trustworthy same-origin destination
PASS content sec-fetch-user - Not sent to non-trustworthy same-origin destination
PASS cursor sec-fetch-user - Not sent to non-trustworthy same-origin destination
PASS list-style-image sec-fetch-user - Not sent to non-trustworthy same-origin destination
TIMEOUT background-image sec-fetch-user - Not sent to non-trustworthy same-site destination Test timed out
FAIL border-image sec-fetch-user - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL content sec-fetch-user - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-user - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-user - Not sent to non-trustworthy same-site destination assert_unreached: Reached unreachable code
PASS background-image sec-fetch-user - Not sent to non-trustworthy cross-site destination
PASS border-image sec-fetch-user - Not sent to non-trustworthy cross-site destination
PASS content sec-fetch-user - Not sent to non-trustworthy cross-site destination
PASS cursor sec-fetch-user - Not sent to non-trustworthy cross-site destination
PASS list-style-image sec-fetch-user - Not sent to non-trustworthy cross-site destination
TIMEOUT background-image sec-fetch-site - HTTPS downgrade (header not sent) Test timed out
FAIL border-image sec-fetch-site - HTTPS downgrade (header not sent) assert_unreached: Reached unreachable code
FAIL content sec-fetch-site - HTTPS downgrade (header not sent) assert_unreached: Reached unreachable code
FAIL cursor sec-fetch-site - HTTPS downgrade (header not sent) assert_unreached: Reached unreachable code
FAIL list-style-image sec-fetch-site - HTTPS downgrade (header not sent) assert_unreached: Reached unreachable code
TIMEOUT background-image sec-fetch-site - HTTPS upgrade Test timed out
PASS border-image sec-fetch-site - HTTPS upgrade
PASS content sec-fetch-site - HTTPS upgrade
PASS cursor sec-fetch-site - HTTPS upgrade
PASS list-style-image sec-fetch-site - HTTPS upgrade
TIMEOUT background-image sec-fetch-site - HTTPS downgrade-upgrade Test timed out
PASS border-image sec-fetch-site - HTTPS downgrade-upgrade
PASS content sec-fetch-site - HTTPS downgrade-upgrade
PASS cursor sec-fetch-site - HTTPS downgrade-upgrade
PASS list-style-image sec-fetch-site - HTTPS downgrade-upgrade
